requirements.txt
absl-py==2.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
computer-vision==0.0.6
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
defusedxml==0.7.1
filelock==3.18.0
flake8==7.2.0
fonttools==4.58.2
fsspec==2025.5.1
ftfy==6.3.1
future==1.0.0
gdown==5.2.0
grpcio==1.73.0
h5py==3.14.0
hf-xet==1.1.3
huggingface-hub==0.33.0
idna==3.10
imageio==2.37.0
isort==4.3.21
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
llvmlite==0.44.0
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mccabe==0.7.0
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
open_clip_torch==2.32.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.2.1
platformdirs==4.3.8
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pycodestyle==2.13.0
pyflakes==3.3.2
pynndescent==0.5.13
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
supervision==0.25.1
sympy==1.14.0
tb-nightly==2.20.0a20250613
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
timm==1.0.15
tokenizers==0.21.1
torch==2.7.1
torchreid==0.2.5
torchvision==0.22.1
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.154
ultralytics-thop==2.0.14
umap==0.1.1
umap-learn==0.5.7
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
yacs==0.1.8
yapf==0.43.0
